pyproject.toml
[project]
name = "gha-multi-gitter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "pyyaml>=6.0.2",
]
